+++ changelog.md
@@ -3,8 +3,8 @@
 # 0.4.0
 - Run commands from newer winrm executor
 - Use latest winrm-fs 0.4.2
-- Wait for inner Powershell process to exit instead of task completion
 - Allow task to run as a service account
+- Provide an artificially long timeout to the task to keep the task from dying after 60 seconds
 
 # 0.3.0
 - [Name Powershell Script and Log Files Uniquely](https://github.com/WinRb/winrm-elevated/pull/6)